Table Size
Table size won't be calculated if the width of the layoutdata has not changed and will not update the calculations if the columns send by the metadata have not changed as well. Beware, if the metadata calculates and is sent before the data has been sent, update data as well afterwards.