Route update
In special edge cases we routed back to the menu even though we had opened another workscreen. In the same vain, we lost workscreens idling in the background as we did not delete screens out of the history correctly, letting the router falsy deleting other routes as unreachable.

E.g. LazyLoad -> First -> Second -> First created a history of
LazyLoad -> First, leaving the second workscreen open in the background but not routable.