Brought back old area calculation
The new way of asking an editor how big it wants to be is simpler, but has a small drawback. The text area will try to keep all 5 rows inside the editor at all times then. Making centering the content impossible, as the text editor will center all 5 rows, not just one with content.