Execute delete screen on successful close screen first, closes FS#3184
The delete screen was always chained with the close screen. But with one important fact: All follow up commands of the close screen were executed before we deleted our screen. This had the unfortunate bug of, when you close a screen and the response is to open it, we do that but delete it afterwards.