fixed routing
sometimes workscreen have been in the history multiple times or have been deleted when they should not have been deleted resulting in flutter errors all around. Changed and unified the routing to workscreen to no depend on the command itself and switched up the logic to check when to route and when not to when exiting a screen while itself another screen gets opened.